apps/file-manager/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
nats-py>=2.6.0
redis>=5.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
pdfminer.six>=20221105
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
zhipuai>=2.0.0
